Refactored Project Structure
Gone from old multi lambda function approach to singular function model
