--- scripts/apigateway/apigateway.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-source params.sh
-
-# Create Register gateway
-aws apigateway create-rest-api --name CMS
