--- scripts/lambda/tests/test-logEvent.json
@@ -1,7 +1,0 @@
-{
-  "Post" : {
-      "Title" : "Hello World!",
-      "Author" : "Adam Campbell",
-      "Content" : "<h1>Hello world!</h1>"
-  }
-}
